+++ tests/test_expressions.py
@@ -3,7 +3,7 @@ def test_simple(mio):
 
 
 def test_complex(mio):
-    assert mio.eval("1 + 2 * 3") == 7
+    assert mio.eval("1 + 2 * 3") == 9
 
 
 def test_grouping(mio):
@@ -15,7 +15,7 @@ def test_assignment(mio):
     assert mio.eval("x") == 1
 
 
-def test_complex_assignment1(mio):
+def test_complex_assignment_expression(mio):
     mio.eval("x = 1")
     assert mio.eval("x") == 1
 
@@ -23,7 +23,7 @@ def test_complex_assignment1(mio):
     assert mio.eval("x") == 2
 
 
-def test_complex_assignment2(mio):
+def test_complex_assignment_attribute(mio):
     mio.eval("Foo = Object clone")
 
     mio.eval("Foo x = 1")