+++ tests/test_parser.py
@@ -3,23 +3,23 @@ from mio.types.string import String
 from mio.parser import parse, tokenize
 
 
-def test_empty_message():
+def test_empty_message(mio):
     assert parse(tokenize("")) is None
 
 
-def test_parens():
+def test_parens(mio):
     chain = parse(tokenize("(1)"))
     assert chain.name == "()"
     assert chain.args[0] == Number(1)
 
 
-def test_parens2():
+def test_parens2(mio):
     chain = parse(tokenize("x(1)"))
     assert chain.name == "x"
     assert chain.args[0] == Number(1)
 
 
-def test_parens3():
+def test_parens3(mio):
     chain = parse(tokenize("x = (1)"))
     assert chain.name == "set"
     assert chain.args[0].name == "x"
@@ -27,20 +27,20 @@ def test_parens3():
     assert chain.args[1].args[0] == Number(1)
 
 
-def test_brackets():
+def test_brackets(mio):
     chain = parse(tokenize("[1]"))
     assert chain.name == "[]"
     assert chain.args[0] == Number(1)
 
 
-def test_brackets2():
+def test_brackets2(mio):
     chain = parse(tokenize("x[1]"))
     assert chain.name == "x"
     assert chain.next.name == "[]"
     assert chain.next.args[0] == Number(1)
 
 
-def test_brackets3():
+def test_brackets3(mio):
     chain = parse(tokenize("x = [1]"))
     assert chain.name == "set"
     assert chain.args[0].name == "x"
@@ -49,20 +49,20 @@ def test_brackets3():
     assert chain.args[1].args[0].args[0] == Number(1)
 
 
-def test_braces():
+def test_braces(mio):
     chain = parse(tokenize("{1}"))
     assert chain.name == "{}"
     assert chain.args[0] == Number(1)
 
 
-def test_braces2():
+def test_braces2(mio):
     chain = parse(tokenize("x{1}"))
     assert chain.name == "x"
     assert chain.next.name == "{}"
     assert chain.next.args[0] == Number(1)
 
 
-def test_braces3():
+def test_braces3(mio):
     chain = parse(tokenize("x = {1}"))
     assert chain.name == "set"
     assert chain.args[0].name == "x"
@@ -71,29 +71,29 @@ def test_braces3():
     assert chain.args[1].args[0].args[0] == Number(1)
 
 
-def test_number_message():
+def test_number_message(mio):
     chain = parse(tokenize("1"))
     assert chain.name == Number(1)
 
 
-def test_string_message():
+def test_string_message(mio):
     chain = parse(tokenize("\"foo\""))
     assert chain.name == String("foo")
 
 
-def test_string_newline():
+def test_string_newline(mio):
     chain = parse(tokenize(r'"\n"'))
     assert chain.name == "\n"
 
 
-def test_simple_assignment():
+def test_simple_assignment(mio):
     chain = parse(tokenize("x = 1"))
     assert chain.name == "set"
     assert chain.args[0].name == "x"
     assert chain.args[1].name == Number(1)
 
 
-def test_multi_assignment():
+def test_multi_assignment(mio):
     chain = parse(tokenize("x = 1\ny = 2"))
     assert chain.name == "set"
     assert chain.args[0].name == "x"
@@ -104,7 +104,7 @@ def test_multi_assignment():
     assert chain.next.next.args[1].name == Number(2)
 
 
-def test_grouped_assignment():
+def test_grouped_assignment(mio):
     chain = parse(tokenize("x = (1)"))
     assert chain.name == "set"
     assert chain.args[0].name == "x"
@@ -112,7 +112,7 @@ def test_grouped_assignment():
     assert chain.args[1].args[0] == Number(1)
 
 
-def test_complex_assignment():
+def test_complex_assignment(mio):
     chain = parse(tokenize("Foo x = 1"))
     assert chain.name == "Foo"
     assert chain.next.name == "set"
@@ -120,7 +120,7 @@ def test_complex_assignment():
     assert chain.next.args[1].name == Number(1)
 
 
-def test_complex_assignment2():
+def test_complex_assignment2(mio):
     chain = parse(tokenize("x = x + 1"))
     assert chain.name == "set"
     assert chain.args[0].name == "x"
@@ -129,41 +129,26 @@ def test_complex_assignment2():
     assert chain.args[1].next.args[0].name == Number(1)
 
 
-def test_chaining():
+def test_chaining(mio):
     chain = parse(tokenize("Foo bar"))
     assert chain.name == "Foo"
     assert chain.next.name == "bar"
 
 
-def test_operators():
+def test_operators(mio):
     chain = parse(tokenize("1 + 2"))
     assert repr(chain) == "1 +(2)"
 
     chain = parse(tokenize("1 + 2 * 3"))
-    assert repr(chain) == "1 +(2 *(3))"
+    assert repr(chain) == "1 +(2) *(3)"
 
 
-def test_operators2():
-    chain = parse(tokenize("foo = method(\n1 +(1)\n)"))
+def test_operators_assignment(mio):
+    chain = parse(tokenize("foo = method(\n1 +1\n)"))
     assert repr(chain) == "set(foo, method(\n 1 +(1) \n))"
 
 
-def test_return():
-    chain = parse(tokenize("foo = method(return 1)"))
-    assert repr(chain) == "set(foo, method(return(1)))"
-
-
-def test_return2():
-    chain = parse(tokenize("foo = method(return 1 + 2)"))
-    assert repr(chain) == "set(foo, method(return(1 +(2))))"
-
-
-def test_return3():
-    chain = parse(tokenize("foo = method(return(1 + 2))"))
-    assert repr(chain) == "set(foo, method(return(1 +(2))))"
-
-
-def test_grouping():
+def test_grouping(mio):
     chain = parse(tokenize("1 + (2 * 3)"))
     assert repr(chain) == "1 +(2 *(3))"